extract Export, ExportMap from hir::def
