+++ src/librustc/hir/def.rs
@@ -1,10 +1,8 @@
 use crate::hir;
-use crate::hir::def_id::{DefId, DefIdMap, CRATE_DEF_INDEX, LOCAL_CRATE};
-use crate::ty;
+use crate::hir::def_id::{DefId, CRATE_DEF_INDEX, LOCAL_CRATE};
 
 use rustc_macros::HashStable;
 use rustc_span::hygiene::MacroKind;
-use rustc_span::Span;
 use syntax::ast;
 use syntax::ast::NodeId;
 
@@ -290,29 +288,6 @@ impl<T> PerNS<Option<T>> {
         use std::iter::once;
 
         once(self.type_ns).chain(once(self.value_ns)).chain(once(self.macro_ns)).filter_map(|it| it)
-    }
-}
-
-/// This is the replacement export map. It maps a module to all of the exports
-/// within.
-pub type ExportMap<Id> = DefIdMap<Vec<Export<Id>>>;
-
-#[derive(Copy, Clone, Debug, RustcEncodable, RustcDecodable, HashStable)]
-pub struct Export<Id> {
-    /// The name of the target.
-    pub ident: ast::Ident,
-    /// The resolution of the target.
-    pub res: Res<Id>,
-    /// The span of the target.
-    pub span: Span,
-    /// The visibility of the export.
-    /// We include non-`pub` exports for hygienic macros that get used from extern crates.
-    pub vis: ty::Visibility,
-}
-
-impl<Id> Export<Id> {
-    pub fn map_id<R>(self, map: impl FnMut(Id) -> R) -> Export<R> {
-        Export { ident: self.ident, res: self.res.map_id(map), span: self.span, vis: self.vis }
     }
 }
