+++ src/librustc_resolve/imports.rs
@@ -13,8 +13,9 @@ use crate::{NameBinding, NameBindingKind, PathResult, PrivacyError, ToNameBindin
 
 use errors::{pluralize, Applicability};
 
-use rustc::hir::def::{self, Export, PartialRes};
+use rustc::hir::def::{self, PartialRes};
 use rustc::hir::def_id::DefId;
+use rustc::hir::exports::Export;
 use rustc::lint::builtin::BuiltinLintDiagnostics;
 use rustc::lint::builtin::{PUB_USE_OF_PRIVATE_EXTERN_CRATE, UNUSED_IMPORTS};
 use rustc::session::DiagnosticMessageId;
